Build-2/requirements.txt
Flask
mysql-connector-python
python-dotenv
Werkzeug
sendgrid